code optimizations need to be completed in emit
